+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.2] - 2019-10-17
+
+- Added
+  - Opt-in deduplication of materials
+- Fixed
+  - Caching of meshes
+
 ## [0.6.1] - 2019-10-14
 
 - Added